+++ src/templates/lenTemplate.ts
@@ -11,7 +11,7 @@ export class LenTemplate extends BaseExpressionTemplate {
     let builder = CompletionItemBuilder
       .create('len', lastComponent)
       .description(`len(expr)`)
-    builder.insertText("len(" + lastComponent + ")")
+    builder.insertText('len(' + lastComponent + ')')
     builder.deleteTextBeforeCursor(position, lastComponent.length + 1)
 
     return builder.build()
@@ -20,16 +20,16 @@ export class LenTemplate extends BaseExpressionTemplate {
 
 export const build = () => new LenTemplate()
 
-function getLastComponent(input: string): string {
-  if (input.length == 0) { return "" }
-  var lastComponent = ""
-  for(var i = 0; i < input.length; i++) {
-      let character = input.substr(input.length - i - 1, 1)
-      if (!character.match(/[a-zA-Z0-9\(\)\[\]\.]/)) {
-          return lastComponent
-      }
-      
-      lastComponent = character + lastComponent
+function getLastComponent (input: string): string {
+  if (input.length === 0) { return '' }
+  let lastComponent = ''
+  for (let i = 0; i < input.length; i++) {
+    let character = input.substr(input.length - i - 1, 1)
+    if (!character.match(/[a-zA-Z0-9\(\)\[\]\.]/)) {
+      return lastComponent
+    }
+
+    lastComponent = character + lastComponent
   }
-  return lastComponent;
+  return lastComponent
 }